+++ server/db/FUN_PRO_VIEWS/29_proc_get_form_data_with_filter.sql
@@ -16,9 +16,9 @@ BEGIN
 		SET @sql = CONCAT('select * from (select record.id as record_id, max(edit_time) as edit_time, u.id as user_id, u.name as user_name, vc.code as user_code, ', @sql, ' , removed as deleted from user_input ui
 		left join users u on (ui.id_users = u.id)
 		left join validation_codes vc on (ui.id_users = vc.id_users)
-		left join sections field on (ui.id_sections = field.id)
-		left join sections form  on (ui.id_section_form = form.id)
+		left join sections field on (ui.id_sections = field.id)		
 		left join user_input_record record  on (ui.id_user_input_record = record.id)
+        LEFT JOIN sections form ON (record.id_sections = form.id)
 		LEFT JOIN sections_fields_translation AS sft_in ON sft_in.id_sections = ui.id_sections AND sft_in.id_fields = 57		
 		where form.id = ', form_id_param, ' group by u.id, u.name, record.id, vc.code, removed) as r where 1=1 ', filter_param);