+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 ### New Feature
  - #487 - keep the `record_id` for the scheduled actions. Add two columns one for `INTERNAL` data type - `id_user_input_record` and one for `EXTERNAL` data type - `id_uploadRows`
  - add extra info from which record a job was scheduled to `scheduledJobs` entry view
- - move `form_id` info from table `user_input` to table `user_input_record`. Normalize data
+ - move `id_section_form` info from table `user_input` to table `user_input_record`. Normalize data. <span style="color:red">**Potential issues with very old SelfHelp versions. Some of the data may not be visible in Admin->Data. It should be visible in the old user input export.**</span>
 
 # v6.14.0
 ### Bugfix